+++ R/internalHelpers.R
@@ -1,4 +1,4 @@
-# Copyright (c) 2023 Apex Resource Management Solution Ltd. (ApexRMS). All rights reserved.
+# Copyright (c) 2024 Apex Resource Management Solution Ltd. (ApexRMS). All rights reserved.
 # MIT License
 #' @include AAAClassDefinitions.R
 NULL
@@ -312,8 +312,7 @@ getFolderData <- function(x) {
   
   # Clean up dataframe names and columns
   names(out) <- sapply(names(out), pascal)
-  colnames(out)[colnames(out) == "IsLite"] ="Published"
-  colnames(out)[colnames(out) == "ID"] ="FolderId"
+  colnames(out)[colnames(out) == "ID"] ="FolderID"
   out <- subset(out, select = -c(X))
   
   if (is(x, "Folder")){
@@ -328,9 +327,12 @@ getFolderData <- function(x) {
 # @param x SyncroSim Library, Project, Scenario, or Folder object.
 # @param folderId integer value of the child Folder ID.
 # @param item string indicating type of the child item ("Folder" or "Scenario")
+# @param item string indicating type of the child item ("Folder" or "Scenario")
 # @return integer corresponding to the parent folder ID or project ID.
 getParentFolderId <- function(x, id, item="Folder") {
+getParentFolderId <- function(x, id, item="Folder") {
   df <- getLibraryStructure(x)
+  childRowInd <- which((df$id == id) & (df$item == item))
   childRowInd <- which((df$id == id) & (df$item == item))
   parentRowInd <- childRowInd - 1
   childRow <- df[childRowInd, ]
@@ -427,6 +429,7 @@ getLibraryStructure <- function(x) {
     i = i + 1
   }
   
+  libStructureDF <- libStructureDF[!is.na(libStructureDF$item),]
   libStructureDF <- libStructureDF[!is.na(libStructureDF$item),]
   return(libStructureDF)
 }